+++ synquill/test/common/test_plain_model_repository.dart
@@ -47,4 +47,7 @@ class TestPlainModelRepository extends SynquillRepositoryBase<PlainModel>
 
   @override
   DatabaseAccessor<GeneratedDatabase> get dao => _dao;
+
+  @override
+  bool get localOnly => false;
 }