+++ synquill/test/common/test_user_repository.dart
@@ -147,4 +147,7 @@ class TestUserRepository extends SynquillRepositoryBase<TestUser> {
 
   /// Get local data count
   int get localCount => _localData.length;
+
+  @override
+  bool get localOnly => false;
 }