+++ synquill/test/integration/sync_queue_smart_update_test.dart
@@ -1117,4 +1117,7 @@ class _TestPlainModelRepository extends SynquillRepositoryBase<PlainModel>
 
   @override
   DatabaseAccessor<GeneratedDatabase> get dao => _dao;
+
+  @override
+  bool get localOnly => false;
 }