+++ synquill/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.5
+* Correctly implemented `localOnly` repositories in mixins.
+> Note: code must be re-generated via build_runner
+
 ## 0.5.4
 * Fixed a bug where `localOnly` repositories might throw an error during `remoteFirst` operations.