+++ synquill_gen/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.5.1
+
+- Generate `localOnly` property for repository implementations.
+> Note: code must be re-generated via build_runner
+
 ## 0.5.0
 
 - Initial version.